comiclib-api/requirements.txt
flask
flask-cors
requests
google-genai
python-dotenv
pydantic<2.12
tenacity<9
